requirements.txt
altair==5.3.0
attrs==23.2.0
bcrypt==4.1.1
blinker==1.8.2
cachetools==5.3.3
captcha==0.5.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
duckdb==1.0.0
extra-streamlit-components==0.1.71
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==1.2.0
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.0
gspread-pandas==3.3.0
httplib2==0.22.0
idna==3.7
jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pandas==1.5.3
pillow==10.3.0
protobuf==5.27.1
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydeck==0.9.1
pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sql-metadata==2.12.0
sqlparse==0.5.0
st-gsheets-connection==0.0.4
streamlit==1.36.0
streamlit-authenticator==0.3.2
tenacity==8.4.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.28.3
watchdog==4.0.1
